first pass at removing side effects when loading files.
