+++ app/views/ui/gallery.jade
@@ -9,7 +9,7 @@
   .user-id
   .street-count
 
-  a.button-like#new-street(href='#', target='_blank', data-i18n='btn.create') Create new street
-  a.button-like#copy-last-street(href='#', target='_blank', data-i18n='btn.copy') Make a copy
+  a.button-like#new-street(href='/new', target='_blank', data-i18n='btn.create') Create new street
+  a.button-like#copy-last-street(href='/copy-last', target='_blank', data-i18n='btn.copy') Make a copy
 
   ul.streets
