+++ assets/scripts/app/blocking_shield.js
@@ -6,22 +6,13 @@
  * showing messages or errors.
  */
 import { msg } from './messages'
-import { goReload } from './routing'
-
-import { blockingCancel, blockingTryAgain } from '../util/fetch_blocking'
 
 const BLOCKING_SHIELD_DARKEN_DELAY = 800
 const BLOCKING_SHIELD_TOO_SLOW_DELAY = 10000
 
-const shieldEl = document.querySelector('#blocking-shield')
-
 let blockingShieldTimerId = -1
 let blockingShieldTooSlowTimerId = -1
 
-// Adds event listeners to the respond to buttons.
-document.querySelector('#blocking-shield-cancel').addEventListener('pointerdown', blockingCancel)
-document.querySelector('#blocking-shield-try-again').addEventListener('pointerdown', blockingTryAgain)
-document.querySelector('#blocking-shield-reload').addEventListener('pointerdown', goReload)
 
 function clearBlockingShieldTimers () {
   window.clearTimeout(blockingShieldTimerId)
@@ -32,6 +23,7 @@ export function showBlockingShield (message = msg('LOADING')) {
   hideBlockingShield()
   clearBlockingShieldTimers()
 
+  const shieldEl = document.querySelector('#blocking-shield')
   shieldEl.querySelector('.message').innerHTML = message
   shieldEl.classList.add('visible')
 
@@ -46,11 +38,13 @@ export function showBlockingShield (message = msg('LOADING')) {
 
 export function darkenBlockingShield (message) {
   clearBlockingShieldTimers()
+  const shieldEl = document.querySelector('#blocking-shield')
   shieldEl.classList.add('darken-immediately')
 }
 
 export function hideBlockingShield () {
   clearBlockingShieldTimers()
+  const shieldEl = document.querySelector('#blocking-shield')
   shieldEl.classList.remove('visible')
   shieldEl.classList.remove('darken')
   shieldEl.classList.remove('darken-immediately')
