+++ assets/scripts/app/debug_info.js
@@ -13,9 +13,6 @@ import { getSettings } from '../users/settings'
 import { registerKeypress, deregisterKeypress } from './keypress'
 import { loseAnyFocus } from './focus'
 
-// Register keyboard input for show (shift-D)
-registerKeypress('shift d', showDebugInfo)
-
 export function showDebugInfo () {
   const debugStreetData = _.cloneDeep(getStreet())
   const debugUndo = _.cloneDeep(getUndoStack())
