+++ assets/scripts/app/keyboard_commands.js
@@ -1,4 +1,3 @@
-import _ from 'lodash'
 
 import { showGallery, hideGallery } from '../gallery/view'
 import { changeBuildingHeight } from '../segments/buildings'
@@ -16,9 +15,6 @@ import { undo, redo } from '../streets/undo_stack'
 import { getSignInData, isSignedIn } from '../users/authentication'
 import { trackEvent } from './event_tracking'
 import { isFocusOnBody } from './focus'
-import { registerKeypress } from './keypress'
-import { msg } from './messages'
-import { showStatusMessage } from './status_message'
 
 export const KEYS = {
   ENTER: 13,
@@ -109,25 +105,3 @@ function getHoveredEl () {
   var el = document.querySelector('.hover')
   return el
 }
-
-// In case anyone tries a save shortcut key out of reflex,
-// we inform the user that it's not necessary.
-registerKeypress('ctrl s', {
-  trackAction: 'Command-S or Ctrl-S save shortcut key pressed'
-}, function () {
-  showStatusMessage(msg('STATUS_NO_NEED_TO_SAVE'))
-})
-
-// Catch-all for the Ctrl-S shortcut from ever trying to
-// save the page contents
-registerKeypress('ctrl s', {
-  preventDefault: true,
-  requireFocusOnBody: false
-}, _.noop)
-
-// Catch-all for the backspace or delete buttons to prevent
-// browsers from going back in history
-registerKeypress(['backspace', 'delete'], {
-  preventDefault: true,
-  requireFocusOnBody: true
-}, _.noop)
