+++ assets/scripts/app/live_update.js
@@ -5,8 +5,6 @@ import { getFetchStreetUrl, unpackServerStreetData } from '../streets/xhr'
 
 const LIVE_UPDATE_DELAY = 5000
 
-const flashEl = document.getElementById('flash')
-
 export function scheduleNextLiveUpdateCheck () {
   window.setTimeout(checkForLiveUpdate, LIVE_UPDATE_DELAY)
 }
@@ -49,6 +47,7 @@ function receiveLiveUpdateStreet (transmission) {
 }
 
 function flash () {
+  const flashEl = document.getElementById('flash')
   flashEl.classList.add('visible')
 
   window.setTimeout(function () {
