+++ assets/scripts/gallery/scroll.js
@@ -5,18 +5,7 @@
 import $ from 'jquery'
 import { getElAbsolutePos } from '../util/helpers'
 
-// This file is in the gallery directory but the same scroll buttons
-// are also added to the palette as well.
-window.addEventListener('stmx:everything_loaded', function () {
-  addScrollButtons(document.querySelector('#palette'))
-  addScrollButtons(document.querySelector('#gallery .streets'))
-})
-
-window.addEventListener('resize', function () {
-  updateScrollButtons()
-})
-
-function addScrollButtons (el) {
+export function addScrollButtons (el) {
   const leftButtonEl = document.createElement('button')
   leftButtonEl.innerHTML = '«'
   leftButtonEl.classList.add('scroll-left')
