+++ assets/scripts/util/fetch_nonblocking.js
@@ -40,10 +40,10 @@ const noConnectionMessage = {
   }
 }
 
-window.addEventListener('stmx:everything_loaded', function () {
+export function attachNonBlockingAjaxListeners() {
   document.querySelector('#no-connection-try-again').addEventListener('pointerdown', nonblockingAjaxTryAgain)
   window.addEventListener('beforeunload', onWindowBeforeUnload)
-})
+}
 
 export function newNonblockingAjaxRequest (request, allowToClosePage, doneFunc, errorFunc, maxRetries) {
   nonblockingAjaxRequestTimer = 0
